fix: EC2에서 스케줄러 작동 안 하던 문제 해결시도
