+++ src/main/java/com/example/barrier_free/domain/favorite/FavoriteService.java
@@ -11,7 +11,6 @@ import java.util.stream.Collectors;
 
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.core.ZSetOperations;
-import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -124,7 +123,7 @@ public class FavoriteService {
 		return "popular:" + year + "-W" + weekNum;
 	}
 
-	private String getLastWeekKey() {
+	public String getLastWeekKey() {
 		LocalDate lastWeek = LocalDate.now().minusWeeks(1);
 		WeekFields weekFields = WeekFields.of(Locale.getDefault());
 		int week = lastWeek.get(weekFields.weekOfWeekBasedYear());
@@ -146,45 +145,11 @@ public class FavoriteService {
 		return favorite;
 	}
 
-	@Scheduled(cron = "0 1 0 ? * TUE")
-	public void cleanupLastWeeklyRankFromRedis() {
-		String prevWeekKey = getLastWeekKey();
-		redisTemplate.delete(prevWeekKey);
-	}
-
-	@Transactional
-	@Scheduled(cron = "0 20 23 ? * MON")
-	public void saveWeeklyRanking() {
-		//지난 주 기준 키를 부르기
-		// popular: 2025 - W2  같은 형식의 키
-		String redisKey = getLastWeekKey();
-
-		//테스트 때문에 이미 있는 경우엔 db에서 지우고 스케쥴링함
-		YearWeek yearWeek = extractYearAndWeek(redisKey);
-
-		Set<ZSetOperations.TypedTuple<Object>> top3 = getTop3FromRedis(redisKey);
-		log.info("[SCHEDULER] 주간 랭킹 저장 시작 - RedisKey: {}", redisKey);
-
-		if (top3 == null || top3.isEmpty()) {
-			log.warn("[SCHEDULER] Redis에 지난주 데이터 없음 - DB 삭제하지 않음");
-			return;
-		}
-		log.info("[SCHEDULER] Redis 랭킹 발견 - 기존 DB 데이터 삭제 후 저장 진행");
-		weeklyRankRepository.deleteByYearAndWeek(yearWeek.getYear(), yearWeek.getWeek());
-
-		List<WeeklyRank> rankings = convertToWeeklyRanks(top3, redisKey);
-		log.info("[SCHEDULER] 저장할 랭킹 수: {}", rankings.size());
-
-		weeklyRankRepository.saveAll(rankings);
-		log.info("[SCHEDULER] 주간 랭킹 저장 완료");
-
-	}
-
-	private Set<ZSetOperations.TypedTuple<Object>> getTop3FromRedis(String key) {
+	public Set<ZSetOperations.TypedTuple<Object>> getTop3FromRedis(String key) {
 		return redisTemplate.opsForZSet().reverseRangeWithScores(key, 0, 2);
 	}
 
-	private List<WeeklyRank> convertToWeeklyRanks(Set<ZSetOperations.TypedTuple<Object>> top3, String redisKey) {
+	public List<WeeklyRank> convertToWeeklyRanks(Set<ZSetOperations.TypedTuple<Object>> top3, String redisKey) {
 		//year 랑 week 나누기
 		YearWeek yearWeek = extractYearAndWeek(redisKey);
 
@@ -204,7 +169,7 @@ public class FavoriteService {
 		return result;
 	}
 
-	private YearWeek extractYearAndWeek(String redisKey) {
+	public YearWeek extractYearAndWeek(String redisKey) {
 		// redisKey - "popular:2025-W27"
 		try {
 			String[] split = redisKey.split(":")[1].split("-W");
